# packages/contracts/openapi/control-api.yaml
openapi: 3.1.0
info:
  title: Pal Control API
  version: 0.4.0
  license:
    name: MIT
    identifier: MIT
  description: >-
    Browser-facing API for Palworld operations. The official Palworld REST API
    and native mod bridge are internal adapters and must not be exposed directly.
    Management operations require a configured administrator API key and are
    additionally restricted to the loopback management boundary. Privileged
    economy and season writes also require a TOTP code and an auditable reason.
    Player-portal sessions use a separate cookie identity and never authorize
    management operations. Non-loopback deployments require HTTPS at a trusted
    reverse proxy in addition to these application controls.
servers:
  - url: http://127.0.0.1:5180/api/v1
    description: Local development
...
paths:
...
  /servers/{serverId}/paldefender/catalog:
    get:
      tags: [PalDefender]
      summary: List the 27 allowlisted PalDefender upstream capabilities
      operationId: getPalDefenderCatalog
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "200":
          description: Static allowlist exposed by this Control API build
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PalDefenderCatalog"
        "404":
          $ref: "#/components/responses/NotFound"
  /servers/{serverId}/paldefender/status:
    get:
      tags: [PalDefender]
      summary: Probe the configured PalDefender adapter and version endpoint
      description: >-
        Always returns a status document for a configured server. A disabled or
        disconnected adapter is represented by connected=false and an error.
      operationId: getPalDefenderStatus
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "200":
          description: PalDefender adapter status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PalDefenderStatus"
        "404":
          $ref: "#/components/responses/NotFound"
  /servers/{serverId}/paldefender/players:
    get:
      tags: [PalDefender]
      summary: List all PalDefender players
      operationId: listPalDefenderPlayers
      x-paldefender-permission: REST.Players.Read
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "200":
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/PalDefenderUnavailable"
        default:
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
  /servers/{serverId}/paldefender/player/{playerIdentifier}:
    get:
      tags: [PalDefender]
      summary: Read one PalDefender player
      operationId: getPalDefenderPlayer
      x-paldefender-permission: REST.Player.Read
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - $ref: "#/components/parameters/PalDefenderPlayerIdentifier"
      responses:
        "200":
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/PalDefenderUnavailable"
        default:
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
  /servers/{serverId}/paldefender/pals/{playerIdentifier}:
    get:
      tags: [PalDefender]
      summary: List a player's Pals through PalDefender
      operationId: listPalDefenderPals
      x-paldefender-permission: REST.Pals.Read
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - $ref: "#/components/parameters/PalDefenderPlayerIdentifier"
      responses:
        "200":
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/PalDefenderUnavailable"
        default:
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
  /servers/{serverId}/paldefender/items/{playerIdentifier}:
    get:
      tags: [PalDefender]
      summary: List a player's items through PalDefender
      operationId: listPalDefenderItems
      x-paldefender-permission: REST.Items.Read
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - $ref: "#/components/parameters/PalDefenderPlayerIdentifier"
      responses:
        "200":
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/PalDefenderUnavailable"
        default:
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
  /servers/{serverId}/paldefender/techs/{playerIdentifier}:
    get:
      tags: [PalDefender]
      summary: List a player's technologies through PalDefender
      operationId: listPalDefenderTechs
      x-paldefender-permission: REST.Techs.Read
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - $ref: "#/components/parameters/PalDefenderPlayerIdentifier"
      responses:
        "200":
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/PalDefenderUnavailable"
        default:
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
  /servers/{serverId}/paldefender/progression/{playerIdentifier}:
    get:
      tags: [PalDefender]
      summary: Read a player's progression through PalDefender
      operationId: getPalDefenderProgression
      x-paldefender-permission: REST.Progression.Read
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - $ref: "#/components/parameters/PalDefenderPlayerIdentifier"
      responses:
        "200":
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/PalDefenderUnavailable"
        default:
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
  /servers/{serverId}/paldefender/guilds:
    get:
      tags: [PalDefender]
      summary: List PalDefender guilds
      operationId: listPalDefenderGuilds
      x-paldefender-permission: REST.Guilds.Read
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "200":
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/PalDefenderUnavailable"
        default:
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
  /servers/{serverId}/paldefender/guild/{guildId}:
    get:
      tags: [PalDefender]
      summary: Read one PalDefender guild and its bases
      operationId: getPalDefenderGuild
      x-paldefender-permission: REST.Guild.Read
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - $ref: "#/components/parameters/PalDefenderGuildId"
      responses:
        "200":
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/PalDefenderUnavailable"
        default:
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
  /servers/{serverId}/paldefender/banlist:
    get:
      tags: [PalDefender]
      summary: Search PalDefender ban records
      operationId: searchPalDefenderBanlist
      x-paldefender-permission: REST.Banlist.Read
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - name: active
          in: query
          schema: { type: boolean }
        - name: entryType
          in: query
          schema: { type: string }
        - name: userId
          in: query
          schema: { type: string }
        - name: ip
          in: query
          schema: { type: string }
        - name: userIP
          in: query
          schema: { type: string }
        - name: issuerType
          in: query
          schema: { type: string }
        - name: issuerName
          in: query
          schema: { type: string }
        - name: issuerIP
          in: query
          schema: { type: string }
        - name: reason
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
        "400":
          description: An unsupported query parameter was supplied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/PalDefenderUnavailable"
        default:
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
...
  /servers/{serverId}/paldefender/give/items/{playerIdentifier}:
    post:
      tags: [PalDefender, Commands]
      summary: Queue a PalDefender item grant
      operationId: givePalDefenderItems
      x-paldefender-permission: REST.Items.Give
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - $ref: "#/components/parameters/PalDefenderPlayerIdentifier"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/PalDefenderCommand"
      responses:
        "200":
          $ref: "#/components/responses/PalDefenderCommandComplete"
        "202":
          $ref: "#/components/responses/PalDefenderCommandAccepted"
        "400":
          $ref: "#/components/responses/PalDefenderBadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/PalDefenderQueueUnavailable"
  /servers/{serverId}/paldefender/give/pals/{playerIdentifier}:
    post:
      tags: [PalDefender, Commands]
      summary: Queue a PalDefender Pal grant
      operationId: givePalDefenderPals
      x-paldefender-permission: REST.Pals.Give
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - $ref: "#/components/parameters/PalDefenderPlayerIdentifier"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/PalDefenderCommand"
      responses:
        "200":
          $ref: "#/components/responses/PalDefenderCommandComplete"
        "202":
          $ref: "#/components/responses/PalDefenderCommandAccepted"
        "400":
          $ref: "#/components/responses/PalDefenderBadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/PalDefenderQueueUnavailable"
...
  /servers/{serverId}/paldefender/Broadcast:
    post:
      tags: [PalDefender, Commands]
      summary: Queue a PalDefender global chat broadcast
      operationId: broadcastPalDefenderMessage
      x-paldefender-permission: REST.Messages.Broadcast
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/PalDefenderCommand"
      responses:
        "200":
          $ref: "#/components/responses/PalDefenderCommandComplete"
        "202":
          $ref: "#/components/responses/PalDefenderCommandAccepted"
        "400":
          $ref: "#/components/responses/PalDefenderBadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/PalDefenderQueueUnavailable"
...
components:
  securitySchemes:
    adminApiKey:
      type: apiKey
      in: header
      name: X-Pal-Admin-Key
      description: >-
        Administrator API key. Configure only a SHA-256 digest at rest; never
        store the plaintext key in repository configuration or logs.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    playerSessionCookie:
      type: apiKey
      in: cookie
      name: pal_player_session
      description: >-
        Opaque player session set only after successful in-game OTP verification.
        The deployment uses an HttpOnly, SameSite=Strict cookie scoped to
        /api/v1/player and enables Secure behind HTTPS. Authenticated player
        identity is derived exclusively from this cookie, never from a userId
        query parameter or business request body.
...
security:
  - adminApiKey: []